requirements.txt
dash
pandas
python-dotenv
requests
upstox-client
gunicorn

